requirements.txt
# Core data processing
pandas>=1.3.0
numpy>=1.21.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Computer Vision and Deep Learning
torch>=1.10.0
torchvision>=0.11.0
opencv-python>=4.5.0

# Data Augmentation
albumentations>=1.3.0

# Configuration Management
PyYAML>=6.0

# Machine Learning Utilities
scikit-learn>=1.0.0

# File and Path Management
pathlib2>=2.3.7

# Logging and Progress
tqdm>=4.62.0

# Optional: For YOLO models (if using Ultralytics)
# ultralytics>=8.0.0

# Optional: For experiment tracking
# wandb>=0.12.0
# tensorboard>=2.8.0 